backend/requirements.txt

fastapi
uvicorn
sqlalchemy
pyodbc
pandas
pypdf2
python-dotenv
numpy
faiss-cpu
pytest
httpx
python-multipart
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.36
langchain-experimental==0.0.58
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.1
psycopg2-binary
python-docx
openpyxl
